Move folder checking code to pkg `dircheck`
