+++ internal/dircheck/checker_test.go
@@ -1,4 +1,4 @@
-package fsprov
+package dircheck
 
 import (
 	"os"
